+++ lua/configs/pause.lua
@@ -1,13 +1,15 @@
 //afkkick default config
 
-DAK.revisions["pause"] = "0.1.211a"
+DAK.revisions["pause"] = "0.1.219a"
 
 local function SetupDefaultConfig()
 	local DefaultConfig = { }
 	DefaultConfig.kPauseChangeDelay = 5
 	DefaultConfig.kPauseMaxPauses = 3
-	DefaultConfig.kPausedReadyNotificationDelay = 10
+	DefaultConfig.kPausedReadyNotificationDelay = 30
 	DefaultConfig.kPausedMaxDuration = 0
+	DefaultConfig.kPauseChatCommands = { "pause" }
+	DefaultConfig.kUnPauseChatCommands = { "unpause" }
 	return DefaultConfig
 end
 
@@ -15,14 +17,14 @@ DAK:RegisterEventHook("PluginDefaultConfigs", {PluginName = "pause", DefaultConf
 
 local function SetupDefaultLanguageStrings()
 	local DefaultLangStrings = { }
-	DefaultLangStrings["PauseResumeMessage"] 					= "Game Resumed.  Team %s has %s pauses remaining"
+	DefaultLangStrings["PauseResumeMessage"] 					= "Game Resumed.  %s have %s pauses remaining"
 	DefaultLangStrings["PausePausedMessage"] 					= "Game Paused."
 	DefaultLangStrings["PauseWarningMessage"] 					= "Game will %s in %.1f seconds."
 	DefaultLangStrings["PauseResumeWarningMessage"] 			= "Game will automatically resume in %.1f seconds."
-	DefaultLangStrings["PausePlayerMessage"] 					= "%s executed a game pause."
-	DefaultLangStrings["PauseTeamReadiedMessage"] 				= "%s readied for Team %s, resuming game."
-	DefaultLangStrings["PauseTeamReadyMessage"] 				= "%s readied for Team %s, waiting for Team %s."
-	DefaultLangStrings["PauseTeamReadyPeriodicMessage"] 		= "Team %s is ready, waiting for Team %s."
+	DefaultLangStrings["PausePlayerMessage"] 					= "%s has paused the game."
+	DefaultLangStrings["PauseTeamReadiedMessage"] 				= "%s readied for %s, resuming game."
+	DefaultLangStrings["PauseTeamReadyMessage"] 				= "%s readied for %s, waiting for the %s."
+	DefaultLangStrings["PauseTeamReadyPeriodicMessage"] 		= "%s are ready, waiting for the %s."
 	DefaultLangStrings["PauseNoTeamReadyMessage"] 				= "No team is ready to resume, type unpause in console to ready for your team."
 	DefaultLangStrings["PauseCancelledMessage"] 				= "Game Pause Cancelled."
 	DefaultLangStrings["PauseTooManyPausesMessage"] 			= "Your team is out of pauses."